// frontend/src/app/learncoreservices/models/tenant-staff-details-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TenantStaffDetailsResponse {
  accountLocked?: boolean;
  avatarUrl?: string;
  createdDate?: string;
  email?: string;
  enabled?: boolean;
  firstname?: string;
  lastname?: string;
  positionName?: string;
  roles?: Array<string>;
  tenantUuid?: string;
  uuid?: string;
}